Harmonize scalac flags and enable Xlint/Werror

+++ sjsonnet/src/sjsonnet/ValVisitor.scala
@@ -42,7 +42,7 @@ class ValVisitor(pos: Position) extends JsVisitor[Val, Val] { self =>
   def visitFloat64StringParts(s: CharSequence, decIndex: Int, expIndex: Int, index: Int): Val =
     Val.Num(pos,
       if (decIndex != -1 || expIndex != -1) s.toString.toDouble
-      else upickle.core.ParseUtils.parseIntegralNum(s, decIndex, expIndex, index)
+      else upickle.core.ParseUtils.parseIntegralNum(s, decIndex, expIndex, index).toDouble
     )
 
   def visitString(s: CharSequence, index: Int): Val = Val.Str(pos, s.toString)
